*: Simplify Server.cpp (#9804)
ref pingcap/tiflash#6233

Simplify Server.cpp

* Reduce parsing `DisaggregatedMode` for multiple times
* Add `Setup.cpp` for setting up the SIMD and jemalloc/mimalloc settings
* Add `TCPServersHolder.cpp`
  * Rename `TcpHttpServersHolder` to `TCPServersHolder` because we never create http server in TiFlash now
  * Merge TCPHandlerFactory into `TCPServersHolder.cpp`
* Remove some useless config inherited from Clickhouse
  * Remove `display_name`, always show "TiFlash" for TCP connections
  * Remove `default_database`, always use "system" for TCP connections
  * Remove `max_concurrent_queries`, no limitation for the number of queries in process list
  * Remove `listen_try` in `TCPServersHolder`
  * Remove `tcp_port_secure` in `TCPServersHolder`
