+++ dbms/src/Server/CMakeLists.txt
@@ -43,8 +43,10 @@ add_library (tiflash-server-lib
     Server.cpp
     StatusFile.cpp
     TCPHandler.cpp
+    TCPServersHolder.cpp
     BgStorageInit.cpp
     Bootstrap.cpp
+    Setup.cpp
     StorageConfigParser.cpp
     UserConfigParser.cpp
     RaftConfigParser.cpp)
