+++ dbms/src/Server/InterruptListener.h
@@ -66,7 +66,7 @@ class InterruptListener
 {
 private:
     bool active;
-    sigset_t sig_set;
+    sigset_t sig_set{};
 
 public:
     InterruptListener()
