+++ dbms/src/Server/Server.h
@@ -62,8 +62,6 @@ private:
     std::unique_ptr<Context> global_context;
 
     ServerInfo server_info;
-
-    class TcpHttpServersHolder;
 };
 
 } // namespace DB
